Rename to upload_to_google_drive_action
