+++ lib/fastlane/plugin/google_drive/actions/google_drive_upload_action.rb
@@ -1,6 +1,7 @@
 module Fastlane
   module Actions
-    class GoogleDriveUploadAction < UploadGoogleDriveAction
+    require 'fastlane/plugin/google_drive/actions/upload_to_google_drive_action'
+    class GoogleDriveUploadAction < UploadToGoogleDriveAction
       #####################################################
       # @!group Documentation
       #####################################################
