+++ spec/upload_to_google_drive_action_spec.rb
@@ -1,6 +1,6 @@
 require 'dotenv/load'
 
-describe Fastlane::Actions::UploadGoogleDriveAction do
+describe Fastlane::Actions::UploadToGoogleDriveAction do
   before(:context) do
     @key_path = File.join(File.dirname(File.dirname(__FILE__)), 'drive_key.json')
     @upload_file = File.join(File.dirname(__FILE__), 'fixtures', 'test_file.txt')
@@ -19,7 +19,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
   it 'raise an error if keyfile does not exists' do
     expect do
       Fastlane::FastFile.new.parse("lane :test do
-        upload_google_drive(drive_keyfile: 'test.json')
+        upload_to_google_drive(drive_keyfile: 'test.json')
       end").runner.execute(:test)
     end.to raise_error(FastlaneCore::Interface::FastlaneError, "Couldn't find config keyfile at path 'test.json'")
   end
@@ -27,7 +27,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
   it 'raise an error if no folder id was given' do
     expect do
       Fastlane::FastFile.new.parse("lane :test do
-        upload_google_drive(drive_keyfile: '#{@key_path}')
+        upload_to_google_drive(drive_keyfile: '#{@key_path}')
       end").runner.execute(:test)
     end.to raise_error(FastlaneCore::Interface::FastlaneError, "No target folder id given, pass using `folder_id: 'some_id'`")
   end
@@ -35,7 +35,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
   it 'raise an error if no upload file was given' do
     expect do
       Fastlane::FastFile.new.parse("lane :test do
-        upload_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: [])
+        upload_to_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: [])
       end").runner.execute(:test)
     end.to raise_error(FastlaneCore::Interface::FastlaneError)
   end
@@ -43,7 +43,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
   it "raise an error if upload file does not exist" do
     expect do
       Fastlane::FastFile.new.parse("lane :test do
-        upload_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: ['nofile'])
+        upload_to_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: ['nofile'])
       end").runner.execute(:test)
     end.to raise_error(FastlaneCore::Interface::FastlaneError, "Couldn't find upload file at path 'nofile'")
   end
@@ -51,7 +51,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
   it "raise an error if folder_id does not exist" do
     expect do
       Fastlane::FastFile.new.parse("lane :test do
-        upload_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: ['#{@upload_file}'])
+        upload_to_google_drive(drive_keyfile: '#{@key_path}', folder_id: 'some_id', upload_files: ['#{@upload_file}'])
       end").runner.execute(:test)
     end.to raise_error(FastlaneCore::Interface::FastlaneError, "File with id 'some_id' not found in Google Drive")
   end
@@ -60,7 +60,7 @@ describe Fastlane::Actions::UploadGoogleDriveAction do
     folder_id = ENV['TEST_UPLOAD_FOLDER_ID']
 
     Fastlane::FastFile.new.parse("lane :test do
-      upload_google_drive(drive_keyfile: '#{@key_path}', folder_id: '#{folder_id}', upload_files: ['#{@upload_file}'])
+      upload_to_google_drive(drive_keyfile: '#{@key_path}', folder_id: '#{folder_id}', upload_files: ['#{@upload_file}'])
     end").runner.execute(:test)
 
     expect(Fastlane::Actions.lane_context[Fastlane::Actions::SharedValues::GDRIVE_UPLOADED_FILE_NAMES]).to eq(['test_file.txt'])
